requirements.txt
PyGithub==2.1.1 
selenium
webdriver-manager